@@ -4,27 +4,33 @@
  *
  * @package SatisPress
  * @author Brady Vercher <brady@blazersix.com>
- * @license GPL-2.0+
+ * @license GPL-2.0-or-later
  *
  * @wordpress-plugin
  * Plugin Name: SatisPress
  * Plugin URI: https://github.com/blazersix/satispress
  * Description: Generate a Composer repository from installed WordPress plugins and themes.
- * Version: 0.2.3
+ * Version: 0.3.0-dev
  * Author: Blazer Six
  * Author URI: http://www.blazersix.com/
- * License: GPL-2.0+
+ * License: GPL-2.0-or-later
  * License URI: http://www.gnu.org/licenses/gpl-2.0.html
  * Text Domain: satispress
  * Domain Path: /languages
+ * GitHub Plugin URI: blazersix/satispress
+ * Requires PHP: 5.6
  */
+
+declare ( strict_types = 1 );
+
+namespace SatisPress;
 
 if ( ! defined( 'SATISPRESS_DIR' ) ) {
 	/**
 	 * Path directory path.
 	 *
 	 * @since 0.2.0
-	 * @type string SATISPRESS_DIR
+	 * @var string SATISPRESS_DIR
 	 */
 	define( 'SATISPRESS_DIR', plugin_dir_path( __FILE__ ) );
 }
@@ -36,33 +42,51 @@ if ( ! defined( 'SATISPRESS_URL' ) ) {
 	 * Includes trailing slash.
 	 *
 	 * @since 0.2.0
-	 * @type string SATISPRESS_URL
+	 * @var string SATISPRESS_URL
 	 */
 	define( 'SATISPRESS_URL', plugin_dir_url( __FILE__ ) );
 }
 
-include( SATISPRESS_DIR . 'includes/functions.php' );
+require SATISPRESS_DIR . 'src/functions.php';
 
+spl_autoload_register( __NAMESPACE__ . '\\satispress_autoloader' );
 /**
  * Autoloader callback.
  *
  * Converts a class name to a file path and requires it if it exists.
  *
  * @since 0.2.0
+ * @link https://www.php-fig.org/psr/psr-4/examples/
  *
  * @param string $class Class name.
  */
-function satispress_autoloader( $class ) {
-	if ( 0 !== strpos( $class, 'SatisPress' ) ) {
+function satispress_autoloader( string $class ) {
+
+	// Project namespace.
+	$prefix = 'SatisPress\\';
+
+	$base_dir = SATISPRESS_DIR . 'src/';
+
+	// Does the class use the namespace prefix?
+	$len = strlen( $prefix );
+
+	if ( 0 !== strncmp( $prefix, $class, $len ) ) {
+		// No, move to the next registered autoloader.
 		return;
 	}
 
-	$file = SATISPRESS_DIR . 'includes/class-' . strtolower( str_replace( '_', '-', $class ) ) . '.php';
+	// Get the relative class name.
+	$relative_class = substr( $class, $len );
+
+	// Replace the namespace prefix with the base directory, replace namespace separators
+	// with directory separators in the relative class name, append with .php.
+	$file = $base_dir . \str_replace( '\\', '/', $relative_class ) . '.php';
 
 	if ( file_exists( $file ) ) {
-		require_once( $file );
+		require_once $file;
 	}
 }
-spl_autoload_register( 'satispress_autoloader' );
 
-SatisPress::instance();
+add_action( 'plugins_loaded', function() {
+	( new Plugin() )->compose();
+} );
