@@ -1,17 +1,30 @@
 <?php
 /**
- * Theme package class.
+ * Theme class
  *
  * @package SatisPress
- * @author Brady Vercher <brady@blazersix.com>
+ * @license GPL-2.0-or-later
  * @since 0.2.0
  */
-class SatisPress_Package_Theme extends SatisPress_Package {
+
+declare ( strict_types = 1 );
+
+namespace SatisPress\PackageType;
+
+use SatisPress\Package;
+use SatisPress\VersionParser;
+
+/**
+ * Theme package class.
+ *
+ * @since 0.2.0
+ */
+class Theme extends Package {
 	/**
 	 * Base path where packages are cached.
 	 *
 	 * @since 0.2.0
-	 * @type string
+	 * @var string
 	 */
 	protected $archive_path;
 
@@ -19,7 +32,7 @@ class SatisPress_Package_Theme extends SatisPress_Package {
 	 * Theme slug.
 	 *
 	 * @since 0.2.0
-	 * @type string
+	 * @var string
 	 */
 	protected $slug;
 
@@ -29,7 +42,7 @@ class SatisPress_Package_Theme extends SatisPress_Package {
 	 * Data cached from wp_get_theme(). Includes theme headers.
 	 *
 	 * @since 0.2.0
-	 * @type array
+	 * @var array
 	 */
 	protected $theme;
 
@@ -38,12 +51,15 @@ class SatisPress_Package_Theme extends SatisPress_Package {
 	 *
 	 * @since 0.2.0
 	 *
-	 * @param string $theme_directory Name of the theme directory.
+	 * @param string        $theme_directory Name of the theme directory.
+	 * @param string        $archive_path    Base path where packages are cached.
+	 * @param VersionParser $version_parser  Version parser.
 	 */
-	public function __construct( $theme_directory, $archive_path ) {
-		$this->slug = $theme_directory;
-		$this->theme = wp_get_theme( $theme_directory );
-		$this->archive_path = trailingslashit( $archive_path );
+	public function __construct( string $theme_directory, string $archive_path, VersionParser $version_parser ) {
+		$this->slug           = $theme_directory;
+		$this->theme          = wp_get_theme( $theme_directory );
+		$this->archive_path   = trailingslashit( $archive_path );
+		$this->version_parser = $version_parser;
 	}
 
 	/**
@@ -53,7 +69,7 @@ class SatisPress_Package_Theme extends SatisPress_Package {
 	 *
 	 * @return boolean
 	 */
-	public function is_installed() {
+	public function is_installed(): bool {
 		return $this->theme->exists();
 	}
 
@@ -64,7 +80,7 @@ class SatisPress_Package_Theme extends SatisPress_Package {
 	 *
 	 * @return string
 	 */
-	public function get_author() {
+	public function get_author(): string {
 		return $this->theme->get( 'Author' );
 	}
 
@@ -75,7 +91,7 @@ class SatisPress_Package_Theme extends SatisPress_Package {
 	 *
 	 * @return string
 	 */
-	public function get_author_uri() {
+	public function get_author_uri(): string {
 		return $this->theme->get( 'AuthorURI' );
 	}
 
@@ -86,7 +102,7 @@ class SatisPress_Package_Theme extends SatisPress_Package {
 	 *
 	 * @return string
 	 */
-	public function get_description() {
+	public function get_description(): string {
 		return $this->theme->get( 'Description' );
 	}
 
@@ -97,7 +113,7 @@ class SatisPress_Package_Theme extends SatisPress_Package {
 	 *
 	 * @return string
 	 */
-	public function get_homepage() {
+	public function get_homepage(): string {
 		return $this->theme->get( 'ThemeURI' );
 	}
 
@@ -108,7 +124,7 @@ class SatisPress_Package_Theme extends SatisPress_Package {
 	 *
 	 * @return string
 	 */
-	public function get_name() {
+	public function get_name(): string {
 		return $this->theme->get( 'Name' );
 	}
 
@@ -121,7 +137,7 @@ class SatisPress_Package_Theme extends SatisPress_Package {
 	 *
 	 * @return string
 	 */
-	public function get_package_root() {
+	public function get_package_root(): string {
 		return get_theme_root( $this->get_slug() );
 	}
 
@@ -132,7 +148,7 @@ class SatisPress_Package_Theme extends SatisPress_Package {
 	 *
 	 * @return string
 	 */
-	public function get_path() {
+	public function get_path(): string {
 		return $this->theme->get_stylesheet_directory();
 	}
 
@@ -146,7 +162,7 @@ class SatisPress_Package_Theme extends SatisPress_Package {
 	 *
 	 * @return string
 	 */
-	public function get_slug() {
+	public function get_slug(): string {
 		return $this->slug;
 	}
 
@@ -157,7 +173,7 @@ class SatisPress_Package_Theme extends SatisPress_Package {
 	 *
 	 * @return string
 	 */
-	public function get_type() {
+	public function get_type(): string {
 		return 'wordpress-theme';
 	}
 
@@ -168,7 +184,7 @@ class SatisPress_Package_Theme extends SatisPress_Package {
 	 *
 	 * @return string
 	 */
-	public function get_version() {
+	public function get_version(): string {
 		return $this->theme->get( 'Version' );
 	}
 }
